chore: rename helper folder to matches
